[Bugfix] Symbols are too delicate to use at boundaries.
In the buggy version, a symbol whose name is "!" was used as a delimiter
of command sequence. However, we have to handle the current package
carefully to make it correctly working on the bounderies like the
outermost interface. Instead of this careful treatment, I decided to use
the name of symbol.

Changing the package of make.lisp to CHIKU.GENSCREENRC is another
approach. In my feeling, it is not so smart. This approach makes the
code more dirtier than the approach in this commit, though the fix in
this commit make the commit dirty.