Body of the MAIN function must be executed in CHIKU.GENSCREENRC package.
This approach achieves 2 things simultaniously:

    1. Retains the package of make.lisp to CL-USER package.
    2. Comparison of hard-coded symbols in CHIKU.GENSCREENRC and read
       symbols works properly.

If 2. was the only puspose, there are some other ways. Use the names of
symbols instead of symbol itself, do not use CHIKU.GENSCREENRC pakcage,
etc. When it comes to 1. however, this is the only way because in
make.lisp I have to load the package.lisp BEFORE changing the package.