Utilize hardstatus instead of echo to avoid the hiding of hardstatus.
The echo command hides the hardstatus line and cannot use string
escapes. Although the higind is useful for outputs the messages that
should not be hidden by any other ouputs, for outputting the key
operation it is not comfortable. The lack of string escapes loads strong
limitation to the contents of output. In order to cope with these
problems, I decided to utilize hardstatus string as message outputs.

The current hard-coded hardstatus string preserves 30 charcters width
for outputs from key bindings by using ``%030=''. The substring from
``%030'' to the end of the string should be identical to the default
hardstatus string that is set by ``hardstatus string'' in .screenrc.

    "%030=|%-w%{=b Mw}%{+u}%{+s}%n %t%{-}%{-}%{-}%+w|%=%m/%d%02c"

Therefore, the setting ``hardstatus string'' should be output from this
converter. And of course, the hardstatus string should not be hard-coded.

Since the screen does not provide any infrastructure to trigger a screen
command when we enter to the state what I'm calling ``raw'' state, we
cannot fix the hardstatus string after after the last key input. Let me
take an example.

    bind -c snormal c eval 'screen' 'hardstatus
    string "%{= }%?%E[snormal] (screen)%:[raw]%?%030=|%-w%{=b
    Mw}%{+u}%{+s}%n %t%{-}%{-}%{-}%+w|%=%m/%d %02c"'

``%E'' toggles ``%?'' into true when the screen is key input waiting
state. Above command changes the hardstatus string to show
``[snormal] (screen)'' and ``[raw]'' during key input waiting state and raw
state respectively, and enter the raw state. Thus, after leaving
``snormal'' state by ``c'' key, if the escape key is input again, then
the ``[snormal] (screen)'' is shown in the preserved 30 character width
space of the hardstatus line.
We can fix the hardstatus string on the next input of escape key by
using ``bindkey''. Suppose that ^z is the escape key. In this case,
follwing line woks:

    bindkey ^z eval 'command' 'hardstatus string "%{= }%?%E[snormal]
    (screen)%:[raw]%?%030=|%-w%{=b Mw}%{+u}%{+s}%n
    %t%{-}%{-}%{-}%+w|%=%m/%d %02c"'

As far as I know, ``bindkey ^z eval 'command''' does not harms at all if
``escape ^z^z'' is set. I thus propose to take the escape key setting as
one of the input to this converter and outputs both of ``escape'' and
``bindkey'' lines.