Symobls must be quoted even in the second argument for KEYBIND macro.
I decided to compel quote a symbol given as the second argument (KEY)
for KEYBIND macro. In other words, the second argument of KEYBIND macro
will be evaluated. The second argument for KEYBIND-COMMON and
MULTIPLE-KEYBINDS macros are also evaluated.

There are 2 reasons behind this decision. The first reason is the case
sensitivity. Upper case keys is normaly used. The second reason is the
insertion mode. In order to define that mode, keys given to KEYBIND
macro must be evaluated.

However, in the screenrc-src file, non-quoted symbols are available as
key designators. And the case is preserved even if the key is not
wrapped by double quotes or vertical bars. This is achieved by changing
the READTABLE-CASE slot temporary. For example,
    {snormal snormal [j focus down]}
is translated into something like:
    (MULTIPLE-KEYBINDS T "snormal" "snormal" "j" ("focus" "down"))