The construction policy of arguments for eval command is changed.
In the new construction, each argument for eval command is non-quoted
string and each argument is composed of double-quoted string. The
arguments are separated by space. Example:
    eval "split"\40"-v" "echo"\40"\"(split -v)\""
In non-quoted string, \40 works as a space. This space is a component of
non-quoted string, not a separator of 2 non-quoted strings. If a user
wants to write a hardstatus string for example, then the user must
express the contents of escaped string as Lisp string literal.
    hardstatus-string "%%\"%c\""
This setting shows one % followed by current time wrapped by double
quotes. The user must not (or, do not have to) escape wrapping double
quotes. On the other hand, the user must escape the % by prepending
additional %, by him/herself. In other words, the user must not write
escape process to make it double-quoted string, but he/she have to write
escape process to make the contents. What screen-keybind assume is
really easy thing: the user have to write contents, not a string that
will be recognized as contents he/she want to express.

How to output one ^ in hardstatus line is a good quiz to understand
above concept. If we do not use eval and directly use hardstatus
command, then following is one of the possible answer:
    hardstatus string "\^\^"
But for binding few commands to a key, we have to use eval. Thus, we
need a string whose contents is same to "\^\^".
    eval "hardstatus"\40"string"\40"\"\\\^\\\^\""
        or
    eval 'hardstatus string "\"\\\^\\\^\""'
        and so on.
Both of " and ^ have to be escaped by prepending one \ in double-quoted
string. The escape for \ is not required in some case, but in this
example it is required to surely put one \ in front of ^.

Previously, each argument was single-quoted string and each component
was non-quoted string or double-quoted string. However, single-quoted
string cannot inlcude a single quote as it's element. Thus I stop to use
single-quoted string, even though it is really good in some other
aspects.