Insert mode is implemented.
The primary purpose of insert mode is to control the state after the
input for some command has been completed. One tipycal example of
commands that invokes the input is colon command. Insert mode sends the
input keys to the input area and move to the desierd state when
Enter,Return or C-g is entered.

Suppose that we bind c to colon command in sn mode and wants to stay at
sn mode after the use of colon command.  If we simply write a binding
that stay at sn mode as follows, it stays at sn mode just after c is
entered.
    bind -c sn c eval "colon" "command"\40"-c"\40"snormal"
As a result, when the input for colon command starts, it is in the sn
mode. Thus as long as the user enters keys that are bound in sn mode,
the key is not entered to the input area (shunning this extra key
operation is the secondary purpose of the insert mode). And, once a key
that is not bound in sn mode is entered, the mode became raw mode and we
have no way to go back to sn mode, the input for input area become
available, though.

The solution is quite easy. If a key that is bound to a command that
invokes the input area is entered, move to sn-insert mode. In the
sn-insert mode, all the key input is sent to the input area by binding
each key to the stuff command, and stays sn-insert mode.
    bind -c sn-insert a eval "stuff a"\40"command -c sn-insert"
But a few special key that announces the complete of input triggers the
transition to sn mode. Enter key, Return key are quite proper for such
key. C-g is a candidate of cancel key because it is a cancel key in
windolist -b and copy mode.
    bind -c sn-insert ^M eval "stuff ^M"\40"command -c sn"
    bind -c sn-insert ^G eval "stuff ^G"\40"command -c sn"

All the states that can be the after state of input must have it's insert
mode. During the output of the keybind settings that is explicitly
written in the input file, such state is collected to the SYMBOL-PLIST of
REQUIRE-INSERT symbol. GEN-INSERT-STATE macro generates the insert mode
settings for the given state.

In order to impelement the insertion mode, START and GOAL arguments are
changed to be evaluated.

The list of commnds that invokes input is hard-coded. I believe this is
not a problem.

Since this is the trial implementation, some codes are not beautiful or
clear enough. Here is the list of fix-required points.
    1. Use of SYMBOL-PLIST to stack states that requires insert mode
    2. Algorithm for detecting the necessity of insert-mode
    3. List of commands which invokes input area