INPUT-INVOKE-P is introduced.
This is still incomplete. For the complete version, a little more
complex system that constructs some combination of insert modes and a
variation of insert mode that allows only one key input are required.

In order to implement INPUT-INVOKE-P in this version, EVALARG source
that is expressed as a list of comitems is desiable on the
KEYBIND-DESC's level. I decided to introduce a class called COMITEMSEQ
and pass instances of this class to KEYBIND-DESC. COMITEMSEQ holds
command and its arguments, the material of evalargs in short.
KEYBIND-DESC will finally construct EVALARGs by using EVALARGIFY
function, the function used to call MAKE-EVALARG, through new
MAKE-EVALARG function.  The judgement about input-requiring commands is
more complex than what I expected. It requires not only the command name
but also what argument is used or how many arguments are given. The
splitting of a EVALARG is quite difficult. In my consideration, it is
impossible to do it correctly in general because any kind of separator
can be included as a content of hardstatus string argument. I therefore
decided to carry out the evalarg-ify on KEYBIND-DESC level by
introducing COMITEMSEQ class.