Insert mode is improved.
There are some commands that requires multiple items to be input like
password command. It allows to cancel the input during the input of
first item or second item. This is implemented as a insert mode named
``layer.'' The digraph command requires another variation of insert
mode. It accepts only 2 times key stroke and allows canceling before or
after the first key stroke. This is implemented as a insert mode named
``transient.''

The cancel key is fixed to C-G and the finish keys are fixed to C-J and
C-M, currently.